refactor: 重写负载需求的方案结构，提取公共函数和变量，top2应用到top1
