+++ src/cpu_common/process_monitor.rs
@@ -1,4 +1,4 @@
-use crate::policy::usage_top2::policy_unname2::UNNAME_TIDS;
+use crate::policy::usage::UNNAME_TIDS;
 use anyhow::{anyhow, Result};
 use flume::{Receiver, Sender};
 use hashbrown::{hash_map::Entry, HashMap};
