+++ src/policy/mod.rs
@@ -1,4 +1,3 @@
 pub mod name_match;
 pub mod pkg_cfg;
-pub mod usage_top1;
-pub mod usage_top2;
+pub mod usage;
